Release v0.4.0: Correlation ID system and position precision fix
Major release with breaking changes:

Breaking Changes:
- Position/FrameData fields changed from f32 to f64 for sub-millimeter precision
- send_command() now returns correlation ID (u64) instead of sequence ID (u32)

New Features:
- Correlation ID system for tracking requests/responses across async boundaries
- Helper functions: send_and_wait_for_completion(), wait_on_correlation_completion()
- High-precision position data (f64) eliminates JSON rounding errors
- SentInstructionInfo broadcast channel for instruction tracking

Bug Fixes:
- Fixed invalid sequence ID errors (RMIT-029, error code 2556957)
  - Web app now sends sequence_id: 0 (driver assigns)
  - Sequence IDs assigned at send time (after priority queue reordering)
  - Driver reads NextSequenceID from controller
- Fixed position display precision loss in UI
- Fixed web app trunk build issues (consolidated entry points)

Documentation:
- Comprehensive migration guide for downstream apps
- Complete release notes with upgrade instructions
- Reorganized documentation structure
- New examples for correlation ID usage patterns

See docs/releases/RELEASE_NOTES_v0.4.0.md for full details.

+++ fanuc_rmi/src/lib.rs
@@ -37,12 +37,12 @@ pub mod protocol {
 #[cfg_attr(feature = "DTO", mirror_dto)]
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub struct FrameData {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
-    pub w: f32,
-    pub p: f32,
-    pub r: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
+    pub w: f64,
+    pub p: f64,
+    pub r: f64,
 }
 
 /// Robot configuration data structure
@@ -125,18 +125,18 @@ impl Default for Configuration {
 #[derive(Serialize, Deserialize, Debug, Clone,Copy, PartialEq)]
 #[serde(rename_all = "PascalCase")]
 pub struct Position {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
-    pub w: f32,
-    pub p: f32,
-    pub r: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
+    pub w: f64,
+    pub p: f64,
+    pub r: f64,
     #[serde(default)]
-    pub ext1: f32,
+    pub ext1: f64,
     #[serde(default)]
-    pub ext2: f32,
+    pub ext2: f64,
     #[serde(default)]
-    pub ext3: f32,
+    pub ext3: f64,
 }
 
 impl Default for Position {
